wip: Clean up a bit
