+++ web/src/components/Cleanup.vue
@@ -12,87 +12,12 @@ import {
 } from '@/utils/constants';
 import { base26Converter } from '@/utils';
 import SaveStatus from '@/components/SaveStatus.vue';
+import DataTable from '@/components/DataTable.vue';
 
-function updateTable(el, binding) {
-  const { dataset, id, activeClasses, setSelection } = binding.value;
-  const colgroup = el.getElementsByTagName('colgroup')[0];
-  const body = el.getElementsByTagName('tbody')[0];
-  const columns = colgroup.children;
-  const rows = body.children;
-  for(let index = 0; index < (columns.length - 1); index += 1) {
-    const col = columns[index + 1]; // Account for 0th being empty
-    col.classList.remove('first', 'last', 'active', 'key', 'group', 'metadata', 'masked', 'measurement');
-    col.classList.add(...activeClasses(index, 'column'));
-    col.classList.add(dataset.column.labels[index]);
-  }
-
-  for(let index = 0; index < (rows.length); index += 1) {
-    const row = rows[index];
-    row.classList.remove('first', 'last', 'active', 'header', 'metadata', 'masked', 'sample');
-    row.classList.add(...activeClasses(index, 'row'));
-    row.classList.add(dataset.row.labels[index]);
-  }
-}
-
-function renderTable(el, binding) {
-  while(el.firstChild) {
-    el.removeChild(el.firstChild);
-  }
-
-  const { dataset, id, activeClasses, setSelection } = binding.value;
-  const thead = document.createElement('thead');
-  const colgroup = document.createElement('colgroup');
-  const tr0 = document.createElement('tr');
-  const th0 = document.createElement('th');
-  const col0 = document.createElement('col');
-  tr0.appendChild(th0);
-  colgroup.appendChild(col0);
-
-  dataset.column.labels.forEach((col, index) => {
-    const coln = document.createElement('col');
-    const thn = document.createElement('th');
-    const span = document.createElement('span');
-    span.innerText = base26Converter(index + 1);
-    thn.onclick = (event) => {
-      setSelection({ key: id, event, axis: 'column', idx: index });
-    }
-    thn.classList.add('control', 'px-2')
-    coln.classList.add(...activeClasses(index, 'column'));
-    coln.classList.add(dataset.column.labels[index]);
-    thn.appendChild(span);
-    tr0.appendChild(thn);
-    colgroup.appendChild(coln);
-  });
-  thead.appendChild(tr0);
-  
-  const tbody = document.createElement('tbody');
-  dataset.sourcerows.forEach((row, index) => {
-    const trn = document.createElement('tr');
-    trn.classList.add(...['datarow'].concat(activeClasses(index, 'row')));
-    trn.classList.add(dataset.row.labels[index]);
-    tbody.appendChild(trn);
-    const td = document.createElement('td');
-    td.innerText = index + 1;
-    td.classList.add('control');
-    td.onclick = (event) => {
-      setSelection({ key: id, event, axis: 'row', idx: index });
-    }
-    trn.appendChild(td);
-    row.forEach((col, idx2) => {
-      const tdn = document.createElement('td');
-      tdn.innerText = col;
-      trn.appendChild(tdn);
-      tdn.classList.add('row');
-    });
-  });
-
-  el.appendChild(colgroup);
-  el.appendChild(thead);
-  el.appendChild(tbody);
-}
 
 export default {
   components: {
+    DataTable,
     SaveStatus,
   },
   props: {
@@ -147,12 +72,6 @@ export default {
       this.imputation.mcar = this.dataset.imputationMCAR;
     },
   },
-  directives: {
-    dataTable: {
-      inserted: renderTable,
-      update: updateTable,
-    }
-  },
   methods: {
     ...mapMutations({ setSelection: SET_SELECTION }),
     base26Converter,
@@ -167,23 +86,6 @@ export default {
         dataset_id: this.id,
         options: this.imputation,
       });
-    },
-    activeClasses(index, axisName) {
-      if (axisName === this.selected.type) {
-        const includes = this.selected.ranges.includes(index);
-        const classList = [];
-        if (includes.member) {
-          classList.push('active');
-        }
-        if (includes.first) {
-          classList.push('first');
-        }
-        if (includes.last) {
-          classList.push('last');
-        }
-        return classList;
-      }
-      return [];
     },
   },
 };
@@ -247,17 +149,10 @@ v-layout.cleanup-wrapper(row)
       v-icon {{ $vuetify.icons.download }}
 
     .overflow-auto
-      table.cleanup-table(v-data-table="this")
+      data-table(v-bind="{ id, dataset, selected }", @setSelection="setSelection")
 </template>
 
 <style lang="scss">
-@mixin masked() {
-  background-color: var(--v-secondary-lighten3);
-  box-shadow: inset 0 0 0 .5px var(--v-secondary-lighten2);
-  font-weight: 300;
-  color: var(--v-secondary-base);
-}
-
 .cleanup-wrapper {
   width: 100%;
 
@@ -299,97 +194,5 @@ v-layout.cleanup-wrapper(row)
       margin: 4px;
     }
   }
-
-  .cleanup-table {
-    border-spacing: 0px;
-    user-select: none;
-    table-layout: fixed;
-    border-collapse: collapse;
-
-    .key, .metadata, .header, .group {
-      color: white;
-      font-weight: 700;
-      text-align: left;
-    }
-    
-    colgroup {
-      col {
-        &.active {
-          background: linear-gradient(0deg,rgba(161, 213, 255, 0.3),rgba(161, 213, 255, 0.3));
-          border-left-color: red !important;
-        }
-
-        &.active.key, &.key {
-          background-color:  var(--v-primary-lighten4);
-        }
-
-        &.active.metadata, &.metadata {
-          background-color:  var(--v-accent2-lighten3);
-        }
-
-        &.active.group, &.group {
-          background-color:   var(--v-accent3-lighten3);
-        }
-      }
-    }
-
-    tr {
-
-      th, td {
-        // border: 1px solid var(--v-secondary-lighten1);
-        padding: 2px 5px;
-        text-align: center;
-        white-space: nowrap;
-
-        &.control {
-          cursor: pointer;
-          font-weight: 300;
-        }
-      }
-
-      &.active td {
-        background: linear-gradient(0deg,rgba(161, 213, 255, 0.3),rgba(161, 213, 255, 0.3));
-
-        &.group,
-        &.key,
-        &.metadata,
-        &.masked {
-          // box-shadow: inset 0 0 0 .5px rgba(161, 213, 255, 0.15) !important;
-        }
-      }
-    }
-
-    tr.datarow {
-      
-      &.header {
-        td {
-          background-color: var(--v-accent-lighten1);
-        }
-      }
-
-      &.metadata {
-        td {
-          background-color: var(--v-accent2-lighten2);
-          // box-shadow: inset 0 0 0 .5px var(--v-accent2-lighten1);
-        }
-      }
-
-      &.header,
-      &.metadata {
-        td.key, td.metadata, td.group {
-          @include masked();
-        }
-      }
-    }
-
-    tr.datarow {
-      &.masked td,
-      td.masked,
-      th.masked {
-        @include masked();
-      }
-    }
-  }
 }
-
 </style>
